+++ jio-jdbc/src/main/java/jio/jdbc/QueryOneStm.java
@@ -18,7 +18,7 @@ import java.util.function.Function;
  * @param <I> Type of the input parameters for the SQL query.
  * @param <O> Type of the objects produced by the result set mapper.
  */
-public final class QueryOneStm<I, O> implements JdbcLambda<I, O> {
+final class QueryOneStm<I, O> implements JdbcLambda<I, O> {
 
 
     private final ResultSetMapper<O> mapper;